requirements.txt
BeautifulSoup==3.2.1
Flask==0.10.1
Flask-SQLAlchemy==0.16
Jinja2==2.7
Markdown==2.3.1
MarkupSafe==0.18
SQLAlchemy==0.8.1
Werkzeug==0.9.1
coaster==0.3.12
distribute==0.6.28
gunicorn==0.17.4
ipython==0.13.2
itsdangerous==0.21
psycopg2==2.5
pytz==2013b
recaptcha-client==1.0.6
semantic-version==2.2.0
simplejson==3.3.0
stevedore==0.8
virtualenv==1.7.2
virtualenv-clone==0.2.4
virtualenvwrapper==4.0
webassets==0.8
wsgiref==0.1.2
